[scheduler] semi-stable search read description
search is returning accurately. there's multiple layers of things that can be adjusted and improved, first-level things like marking up links, then there's second level things like office hours shouldnt be able to conflict with teaching schedule. third level like the redundnecy in the functions -- for instance the parsing functions started getting that eloquence but it was based on the avaliablity of the long function for list_master_person
in the meantime, the immediate things to "finish" are fixing the available now and the js trigger on the front-end dashboard for hiding additiional fields